Module wasmer_cli::c_gen::object_file_header

Generate a header file for the object file produced by the ObjectFile engine.

Constants

HELPER_FUNCTIONS
Helper functions to simplify the usage of the object file engine.

Functions

generate_header_file
Generate the header file that goes with the generated object file.