Module wasmer_cli::c_gen::staticlib_header

Generate a header file for the static object file produced.

Functions

gen_helper_functions 🔒

Helper functions to simplify the usage of the static artifact.

generate_header_file

Generate the header file that goes with the generated object file.