wasmer_cli::commands::app::create

Struct CmdAppCreate

pub struct CmdAppCreate {

    pub template: Option<String>,
    pub package: Option<String>,
    pub use_local_manifest: bool,
    pub deploy_app: bool,
    pub no_validate: bool,
    pub non_interactive: bool,
    pub offline: bool,
    pub owner: Option<String>,
    pub app_name: Option<String>,
    pub app_dir_path: Option<PathBuf>,
    pub no_wait: bool,
    pub env: WasmerEnv,
    pub fmt: ItemFormatOpts,
    pub new_package_name: Option<String>,
    pub quiet: bool,

}

Create a new Edge app.

Fields
template: Option<String>

A reference to the template to use.

It can be either an URL to a github repository - like https://github.com/wasmer-examples/php-wasmer-starter - or the name of a template that will be searched for in the selected registry, like astro-starter.

package: Option<String>

Name of the package to use.

use_local_manifest: bool

Whether or not to search (and use) a local manifest.

deploy_app: bool

Whether or not to deploy the application once it is created.

If selected, this might entail the step of publishing the package related to the application. By default, the application is not deployed and the package is not published.

no_validate: bool

Skip local schema validation.

non_interactive: bool

Do not prompt for user input.

offline: bool

Do not interact with any APIs.

owner: Option<String>

The owner of the app.

app_name: Option<String>

The name of the app (can be changed later)

app_dir_path: Option<PathBuf>

The path to the directory where the config file for the application will be written to.

no_wait: bool

Do not wait for the app to become reachable if deployed.

env: WasmerEnvfmt: ItemFormatOptsnew_package_name: Option<String>

Name to use when creating a new package from a template.

quiet: bool

Don’t print any message.

Implementations

impl CmdAppCreate

fn get_app_config(&self, owner: &str, name: &str, package: &str) -> AppConfigV1

async fn get_app_name(&self) -> Result<String>

async fn get_owner(&self, client: Option<&WasmerClient>) -> Result<String>

async fn get_output_dir(&self, app_name: &str) -> Result<PathBuf>

async fn create_from_local_manifest( &self, owner: &str, app_name: &str, ) -> Result<bool>

async fn create_from_package( &self, client: Option<&WasmerClient>, owner: &str, app_name: &str, ) -> Result<bool>

fn persist_in_cache<S: Serialize>(path: &Path, data: &S) -> Result<(), Error>

async fn fetch_templates_cached( client: &WasmerClient, cache_dir: &Path, language: &str, ) -> Result<Vec<AppTemplate>, Error>

Tries to retrieve templates from a local file cache. Fetches the templates from the backend if the file doesn’t exist, can’t be loaded, or is older then the max age,

fn load_cached<D: DeserializeOwned>(path: &Path) -> Result<(D, Duration), Error>

Load cached data from a file.

Returns an error if the cache file is older than the max age.

async fn fetch_template_languages_cached( client: &WasmerClient, cache_dir: &Path, ) -> Result<Vec<TemplateLanguage>>

async fn get_template_url( &self, client: &WasmerClient, ) -> Result<(Url, Option<PathBuf>)>

async fn create_from_template( &self, client: Option<&WasmerClient>, owner: &str, app_name: &str, ) -> Result<bool>

async fn try_deploy( &self, owner: &str, app_name: &str, path: Option<&Path>, build_remote: bool, skip_prompt: bool, ) -> Result<()>