wasmer_cli::commands::auth::login

Module auth_server

Structs

BrowserAuthContext ğŸ”’

A utility struct used to manage the local server for browser-based authorization.

ValidatedNonceOutput ğŸ”’

Payload from the frontend after the user has authenticated.

Enums

TokenStatus ğŸ”’

Payload from the frontend after the user has authenticated.

Functions

handle_post_save_token ğŸ”’

handle_unknown_method ğŸ”’

preflight ğŸ”’

service_router ğŸ”’

setup_listener ğŸ”’