Struct wasmer_cli::commands::Run

pub struct Run {
    disable_cache: bool,
    path: PathBuf,
    invoke: Option<String>,
    command_name: Option<String>,
    cache_key: Option<String>,
    store: StoreOptions,
    wasi: Wasi,
    verbose: u8,
    args: Vec<String>,
}

The options for the wasmer run subcommand

Fields
disable_cache: bool

Disable the cache

path: PathBuf

File to run

invoke: Option<String>

Invoke a specified function

command_name: Option<String>

The command name is a string that will override the first argument passed to the wasm program. This is used in wapm to provide nicer output in help commands and error messages of the running wasm program

cache_key: Option<String>

A prehashed string, used to speed up start times by avoiding hashing the wasm module. If the specified hash is not found, Wasmer will hash the module as if no cache-key argument was passed.

store: StoreOptionswasi: Wasiverbose: u8args: Vec<String>

Application arguments

Implementations

impl Run

pub fn execute(&self) -> Result<()>

Execute the run command

fn inner_execute(&self) -> Result<()>

fn get_module(&self) -> Result<Module>

fn get_module_from_cache(
&self,
store: &Store,
contents: &[u8],
engine_type: &EngineType,
compiler_type: &CompilerType
) -> Result<Module>

fn get_cache(
&self,
engine_type: &EngineType,
compiler_type: &CompilerType
) -> Result<FileSystemCache>

Get the Compiler Filesystem cache

fn try_find_function(
&self,
instance: &Instance,
name: &str,
args: &[String]
) -> Result<Function>

fn invoke_function(
&self,
instance: &Instance,
invoke: &str,
args: &[String]
) -> Result<Box<[Val]>>