wasmer_cli::utils::prompts

Function prompt_for_namespace

pub fn prompt_for_namespace(
    message: &str,
    default: Option<&str>,
    user: Option<&UserWithNamespaces>,
) -> Result<String, Error>

Prompt for a namespace.

Will either show a select with all available namespaces based on the user argument, or present a basic text input.

The username will be included as an option.