wasmer_cli::utils::prompts

Function prompt_for_package

pub async fn prompt_for_package(
    message: &str,
    default: Option<&str>,
    check: Option<PackageCheckMode>,
    client: Option<&WasmerClient>,
) -> Result<(NamedPackageIdent, Option<Package>), Error>

Ask for a package name.

Will continue looping until the user provides a valid name.

If an API is provided, will check if the package exists.