Crate wasmer_cli

The Wasmer binary lib

Modules

backend 🔒

Common module with common used structures across different commands.

c_gen 🔒

A convenient little abstraction for building up C expressions and generating simple C code.

commands 🔒

The commands available in the Wasmer binary.

common 🔒

Common module with common used structures across different commands.

config 🔒

edge_config 🔒

error 🔒

Implements PretyError to print pretty errors in the CLI (when they happen)

logging 🔒

Logging functions for the debug feature.

opts 🔒

types 🔒

utils 🔒

Utility functions for the WebAssembly module

Macros

warning

A macro that prints a warning with nice colors

Constants

VERSION 🔒

Version number of this crate.

Functions

run_cli

Run the Wasmer CLI.