Crate wasmer_cli

The Wasmer binary lib

Modules

c_gen

A convenient little abstraction for building up C expressions and generating simple C code.

cli

The logic for the Wasmer CLI tool.

commands

The commands available in the Wasmer binary.

common

Common module with common used structures across different commands.

error

Implements PretyError to print pretty errors in the CLI (when they happen)

logging

Logging functions for the debug feature.

package_source

Module for parsing and installing packages

store

Common module with common used structures across different commands.

suggestions

This file provides suggestions for the user, to help them on the usage of WebAssembly

utils

Utility functions for the WebAssembly module

Macros

warning

A macro that prints a warning with nice colors

Constants

VERSION

Version number of this crate.