Module wasmer_cli::c_gen

A convenient little abstraction for building up C expressions and generating simple C code.

Modules

staticlib_header
Generate a header file for the object file produced by the Staticlib engine.

Enums

CStatement
A statement in the C programming language. This may not be exact to what an AST would look like or what the C standard says about the C language, it's simply a structed way to organize data for generating C code.

CType
A Type in the C language.

Functions

generate_c
Generate C source code from some CStatements into a String.

Type Definitions

CIdent
An identifier in C.