Enum wasmer_cli::c_gen::CType

pub enum CType {
    Void,
    PointerTo {
        is_const: bool,
        inner: Box<CType>,
    },
    U8,
    U16,
    U32,
    U64,
    USize,
    I8,
    I16,
    I32,
    I64,
    ISize,
    Function {
        arguments: Vec<CType>,
        return_value: Option<Box<CType>>,
    },
    Array {
        inner: Box<CType>,
    },
    TypeDef(String),
}

A Type in the C language.

Variants

Void

C void type.

PointerTo

A pointer to some other type.

Fields of PointerTo

is_const: bool

Whether the pointer is const.

inner: Box<CType>

The type that the pointer points to.

U8

C 8 bit unsigned integer type.

U16

C 16 bit unsigned integer type.

U32

C 32 bit unsigned integer type.

U64

C 64 bit unsigned integer type.

USize

C pointer sized unsigned integer type.

I8

C 8 bit signed integer type.

I16

C 16 bit signed integer type.

I32

C 32 bit signed integer type.

I64

C 64 bit signed integer type.

ISize

C pointer sized signed integer type.

Function

A function or function pointer.

Fields of Function

arguments: Vec<CType>

The arguments the function takes.

return_value: Option<Box<CType>>

The return value if it has one

None is equivalent to Some(Box(Ctype::Void)).

Array

C constant array.

Fields of Array

inner: Box<CType>

The type of the array.

TypeDef(String)

A user defined type.

Implementations

impl CType

pub fn void_ptr() -> Self

Convenience function to get a mutable void pointer type.

pub fn const_void_ptr() -> Self

Convenience function to get a const void pointer type.

fn generate_c(&self, w: &mut String)

Generate the C source code for a type into the given String.

fn generate_c_with_name(&self, name: &str, w: &mut String)

Generate the C source code for a type with a nameinto the given String.