Module wasmer_cli::c_gen::staticlib_header

Generate a header file for the object file produced by the Staticlib engine.

Constants

HELPER_FUNCTIONS

Helper functions to simplify the usage of the Staticlib engine.

Functions

generate_header_file

Generate the header file that goes with the generated object file.