Module wasmer_cli::commands

The commands available in the Wasmer binary.

Modules

binfmt 🔒

cache 🔒

compile 🔒

config 🔒

create_exe 🔒

Create a standalone native executable for a given Wasm file.

create_obj 🔒

Create a standalone native executable for a given Wasm file.

inspect 🔒

list 🔒

run 🔒

self_update 🔒

When wasmer self-update is executed, this is what gets executed

validate 🔒

wast 🔒

Runs a .wast WebAssembly test suites

Structs

Binfmt

Unregister and/or register wasmer as binfmt interpreter

Compile

The options for the wasmer compile subcommand

Config

The options for the wasmer config subcommand

CreateExe

The options for the wasmer create-exe subcommand

CreateObj

The options for the wasmer create-exe subcommand

Inspect

The options for the wasmer validate subcommand

List

Subcommand for listing packages

Run

The options for the wasmer run subcommand

RunWithoutFile

Same as wasmer run, but without the required path argument (injected previously)

SelfUpdate

The options for the wasmer self-update subcommand

Validate

The options for the wasmer validate subcommand

Wast

The options for the wasmer wast subcommand

Enums

Cache

The options for the wasmer cache subcommand

ObjectFormat

The kind of object format to emit.

Type Definitions

PrefixerFn

The prefixer returns the a String to prefix each of the functions in the static object generated by the so we can assure no collisions.