wasmer_cli

Module commands

The commands available in the Wasmer binary.

Re-exports

pub use self::run::Run;

pub use binfmt::*;

pub use compile::*;

pub use create_exe::*;

pub use wast::*;

pub use create_obj::*;

pub use gen_c_header::*;

pub use self::journal::*;

pub use self::add::*;

pub use self::auth::*;

pub use self::cache::*;

pub use self::config::*;

pub use self::container::*;

pub use self::init::*;

pub use self::inspect::*;

pub use self::package::*;

pub use self::publish::*;

pub use self::self_update::*;

pub use self::validate::*;

Modules

add ğŸ”’

app ğŸ”’

Edge app commands.

auth ğŸ”’

binfmt ğŸ”’

cache ğŸ”’

compile ğŸ”’

config ğŸ”’

connect ğŸ”’

container ğŸ”’

create_exe ğŸ”’

Create a standalone native executable for a given Wasm file.

create_obj ğŸ”’

Create a standalone native executable for a given Wasm file.

domain ğŸ”’

gen_c_header ğŸ”’

gen_completions ğŸ”’

gen_manpage ğŸ”’

init ğŸ”’

inspect ğŸ”’

journal ğŸ”’

namespace ğŸ”’

package ğŸ”’

run ğŸ”’

self_update ğŸ”’

ssh

Edge SSH command.

validate ğŸ”’

wast ğŸ”’

Runs a .wast WebAssembly test suites

Structs

WasmerCmd

Command-line arguments for the Wasmer CLI.

Enums

Cmd ğŸ”’

The options for the wasmer Command Line Interface

Traits

AsyncCliCommand ğŸ”’

An executable CLI command that runs in an async context.

CliCommand ğŸ”’

An executable CLI command.

Functions

is_binfmt_interpreter ğŸ”’

print_version ğŸ”’