Module wasmer_cli::commands::config

Structs

BoolString

“true” or “false” for handling input in the CLI

Config

The options for the wasmer config subcommand: wasmer config get --OPTION or wasmer config set [FLAG]

Flags

Normal configuration

SetProxyUrl

Set if a proxy URL should be used

SetRegistryToken

Set or change the token for the current active registry

SetRegistryUrl

Set the current active registry URL

SetTelemetryEnabled

Set if telemetry is enabled

SetUpdateNotificationsEnabled

Set if update notifications are enabled

Enums

GetOrSet

Subcommand for wasmer config set

RetrievableConfigField

Subcommand for wasmer config get

StorableConfigField

Setting that can be stored in the wasmer config