Module wasmer_cli::commands::create_exe

Create a standalone native executable for a given Wasm file.

Modules

http_fetch 🔒

Structs

CreateExe

The options for the wasmer create-exe subcommand

CrossCompile 🔒

CrossCompileSetup 🔒

LinkCode 🔒

Data used to run a linking command for generated artifacts.

Constants

WASMER_MAIN_C_SOURCE 🔒

WASMER_STATIC_MAIN_C_SOURCE 🔒

Functions

find_zig_binary 🔒

get_libwasmer_cache_path 🔒

path to library tarball cache dir

get_libwasmer_path 🔒

path to the static libwasmer

get_wasmer_dir 🔒

get_wasmer_include_directory 🔒

get_zig_exe_str 🔒

run_c_compile 🔒

Compile the C code.

triple_to_zig_triple 🔒

untar 🔒

Type Definitions

PrefixerFn

The prefixer returns the a String to prefix each of the functions in the static object generated by the so we can assure no collisions.