wasmer_cli::utils
Module prompts

Enums

PackageCheckMode

Defines how to check for a package.

Functions

prompt_for_app_ident

Ask a user for an application name.

prompt_for_ident

prompt_for_namespace

Prompt for a namespace.

prompt_for_package

Ask for a package name.

prompt_for_package_ident

Ask a user for a package name.

prompt_for_package_version

Ask a user for a package version.

prompt_new_app_alias

Prompt for an app name. If an api provided, will check if an app with the givne alias already exists.

prompt_new_app_name

Prompt for an app name. If an api provided, will check if an app with the givne alias already exists.